Title: 2개씩 그룹짓기 2, Time: 432ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 250112/2개씩 그룹짓기 2/group-of-pairs-2.java
@@ -8,7 +8,7 @@ public class Main {
         List<Long> lst = new ArrayList<>();
 
         for (int i = 0; i < n; i++)
-            lst.add(sc.nextInt());
+            lst.add(sc.nextLong());
 
         Collections.sort(lst);
